Refactor expression parse tests
Refs #3741

* Move unary_ops and binary_ops parse tests into appropriate expression
  tests
