+++ src/webgpu/listing_meta.json
@@ -1971,6 +1971,7 @@
   "webgpu:shader,validation,expression,binary,div_rem:invalid_type_with_itself:*": { "subcaseMS": 38.059 },
   "webgpu:shader,validation,expression,binary,div_rem:scalar_vector:*": { "subcaseMS": 743.721 },
   "webgpu:shader,validation,expression,binary,div_rem:scalar_vector_out_of_range:*": { "subcaseMS": 650.727 },
+  "webgpu:shader,validation,expression,binary,parse:all:*": { "subcaseMS": 527.287 },
   "webgpu:shader,validation,expression,call,builtin,abs:parameters:*": { "subcaseMS": 10.133 },
   "webgpu:shader,validation,expression,call,builtin,abs:values:*": { "subcaseMS": 0.391 },
   "webgpu:shader,validation,expression,call,builtin,acos:integer_argument:*": { "subcaseMS": 1.512 },
@@ -2444,6 +2445,7 @@
   "webgpu:shader,validation,expression,unary,bitwise_complement:invalid_types:*": { "subcaseMS": 7.564 },
   "webgpu:shader,validation,expression,unary,bitwise_complement:scalar_vector:*": { "subcaseMS": 73.852 },
   "webgpu:shader,validation,expression,unary,logical_negation:invalid_types:*": { "subcaseMS": 7.100 },
+  "webgpu:shader,validation,expression,unary,logical_negation:parse:*": { "subcaseMS": 4.927 },
   "webgpu:shader,validation,expression,unary,logical_negation:scalar_vector:*": { "subcaseMS": 84.680 },
   "webgpu:shader,validation,extension,pointer_composite_access:deref:*": { "subcaseMS": 0.000 },
   "webgpu:shader,validation,extension,pointer_composite_access:pointer:*": { "subcaseMS": 0.000 },
@@ -2491,7 +2493,6 @@
   "webgpu:shader,validation,functions,restrictions:parens_required:*": { "subcaseMS": 760.897 },
   "webgpu:shader,validation,functions,restrictions:vertex_returns_position:*": { "subcaseMS": 1.201 },
   "webgpu:shader,validation,parse,attribute:expressions:*": { "subcaseMS": 1.410 },
-  "webgpu:shader,validation,parse,binary_ops:all:*": { "subcaseMS": 1.301 },
   "webgpu:shader,validation,parse,blankspace:blankspace:*": { "subcaseMS": 1.391 },
   "webgpu:shader,validation,parse,blankspace:bom:*": { "subcaseMS": 1.101 },
   "webgpu:shader,validation,parse,blankspace:null_characters:*": { "subcaseMS": 3.217 },
@@ -2597,7 +2598,6 @@
   "webgpu:shader,validation,parse,source:empty:*": { "subcaseMS": 1.101 },
   "webgpu:shader,validation,parse,source:invalid_source:*": { "subcaseMS": 1.100 },
   "webgpu:shader,validation,parse,source:valid_source:*": { "subcaseMS": 1.101 },
-  "webgpu:shader,validation,parse,unary_ops:all:*": { "subcaseMS": 1.000 },
   "webgpu:shader,validation,shader_io,align:multi_align:*": { "subcaseMS": 3.232 },
   "webgpu:shader,validation,shader_io,align:parsing:*": { "subcaseMS": 368.840 },
   "webgpu:shader,validation,shader_io,align:placement:*": { "subcaseMS": 20.963 },
