+++ src/webgpu/shader/validation/expression/unary/logical_negation.spec.ts
@@ -112,3 +112,44 @@ fn main() {
 
     t.expectCompileResult(t.params.control, code);
   });
+
+const kTests = {
+  not_bool_literal: {
+    src: 'let a = !true;',
+    pass: true,
+  },
+  not_bool_expr: {
+    src: `let a = !(1 == 2);`,
+    pass: true,
+  },
+  not_not_bool_literal: {
+    src: 'let a = !!true;',
+    pass: true,
+  },
+  not_not_bool_expr: {
+    src: `let a = !!(1 == 2);`,
+    pass: true,
+  },
+  not_int_literal: {
+    src: `let a = !42;`,
+    pass: false,
+  },
+  not_int_expr: {
+    src: `let a = !(40 + 2);`,
+    pass: false,
+  },
+};
+
+g.test('parse')
+  .desc('Test that unary operators are parsed correctly')
+  .params(u => u.combine('stmt', keysOf(kTests)))
+  .fn(t => {
+    const code = `
+@vertex
+fn vtx() -> @builtin(position) vec4f {
+  ${kTests[t.params.stmt].src}
+  return vec4f(1);
+}
+    `;
+    t.expectCompileResult(kTests[t.params.stmt].pass, code);
+  });
